+++ src/quantum/index.jsx
@@ -4,7 +4,7 @@ import { parse } from 'query-string';
 import Grid from '@material-ui/core/Grid/Grid';
 import DashboardPage from '../components/DashboardPage';
 import PerfherderWidget from './perfherder';
-import { frum, toPairs } from '../vendor/queryOps';
+import { fluent, toPairs } from '../vendor/fluent';
 import { FromQueryString } from '../vendor/convert';
 import TelemetryContainer from '../telemetry/graph';
 import {
@@ -75,7 +75,7 @@ export default class QuantumIndex extends React.Component {
       {
         title: 'Page Load tests (TP6)',
         more: `/quantum/tp6?bits=${bits}&test=loadtime`,
-        rows: frum(TP6_PAGES)
+        rows: fluent(TP6_PAGES)
           .where({ bits })
           .groupBy('title')
           .map((series, title) => (
@@ -567,7 +567,7 @@ export default class QuantumIndex extends React.Component {
     if (full) {
       return (
         <DashboardPage title="Quantum" subtitle="Release Criteria Report">
-          {frum(reduced).limit(2)}
+          {fluent(reduced).limit(2)}
           <h2 key="moreData">
             {'More data on'}
             <strong>https://health.graphics/quantum</strong>
