+++ src/utils/perfherder/subbenchmarks.jsx
@@ -1,6 +1,6 @@
 /* global fetch */
 import percentile from 'aggregatejs/percentile';
-import { frum, toPairs } from '../../vendor/queryOps';
+import { fluent, toPairs } from '../../vendor/fluent';
 import { toQueryString } from '../../vendor/convert';
 
 const TREEHERDER = 'https://treeherder.mozilla.org';
@@ -23,7 +23,7 @@ const parentInfo = async ({ suite, platform, framework, option }) => {
   ]);
   // Create a structure with only jobs matching the suite, make
   // option_collection_hash be the key and track the signatureHash as a property
-  const optionHashes = frum(options)
+  const optionHashes = fluent(options)
     .where({ 'options.0.name': option })
     .select('option_collection_hash')
     .toArray();
