+++ src/views/QuantumTP6m/index.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { withStyles } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import Grid from '@material-ui/core/Grid';
-import { frum } from '../../vendor/queryOps';
+import { fluent } from '../../vendor/fluent';
 import { TP6_TESTS, TP6M_PAGES, PLATFORMS } from '../../quantum/config';
 import { withNavigation } from '../../vendor/utils/navigation';
 import Picker from '../../vendor/utils/navigation/Picker';
@@ -23,7 +23,7 @@ const styles = {
 class TP6M extends React.Component {
   render() {
     const { classes, navigation, test, platform } = this.props;
-    const subtitle = frum(TP6_TESTS)
+    const subtitle = fluent(TP6_TESTS)
       .where({ id: test })
       .first().label;
 
@@ -32,7 +32,7 @@ class TP6M extends React.Component {
         <DashboardPage key={subtitle} title="TP6 Mobile" subtitle={subtitle}>
           {navigation}
           <Grid container spacing={24}>
-            {frum(TP6M_PAGES)
+            {fluent(TP6M_PAGES)
               .where({ platform })
               .groupBy('title')
               .map((series, title) => (
@@ -43,7 +43,7 @@ class TP6M extends React.Component {
                   className={classes.chart}>
                   <PerfherderGraphContainer
                     title={title}
-                    series={frum(series)
+                    series={fluent(series)
                       .sortBy(['browser'])
                       .reverse()
                       .map(s => ({
@@ -83,7 +83,7 @@ const nav = [
     id: 'platform',
     label: 'Platform',
     defaultValue: 'android-hw-g5-7-0-arm7-api-16',
-    options: frum(PLATFORMS)
+    options: fluent(PLATFORMS)
       .where({ browser: 'geckoview' })
       .select({ id: 'platform', label: 'label' })
       .toArray(),
